fix: bump type for `DiskSize` to be 64-bit
Otherwise we're bound with 4GiB partititions.

Discovered by @Unix4ever.

+++ pkg/machinery/config/provider.go
@@ -53,7 +53,7 @@ type Disk interface {
 
 // Partition represents the options for a device partition.
 type Partition interface {
-	Size() uint
+	Size() uint64
 	MountPoint() string
 }
